Move config.js to root
